Selectors, resolving merge errors, better class names, removing comment reducer
