--- client/coral-embed-stream/src/constants/comment.js
@@ -1,2 +1,0 @@
-export const ADD_COMMENT_CLASSNAME = 'ADD_COMMENT_CLASSNAME';
-export const REMOVE_COMMENT_CLASSNAME = 'REMOVE_COMMENT_CLASSNAME';
