+++ client/coral-embed-stream/src/constants/stream.js
@@ -1,3 +1,5 @@
 export const SET_ACTIVE_REPLY_BOX = 'SET_ACTIVE_REPLY_BOX';
 export const ADDTL_COMMENTS_ON_LOAD_MORE = 10;
 export const VIEW_ALL_COMMENTS = 'VIEW_ALL_COMMENTS';
+export const ADD_COMMENT_CLASSNAME = 'ADD_COMMENT_CLASSNAME';
+export const REMOVE_COMMENT_CLASSNAME = 'REMOVE_COMMENT_CLASSNAME';
