+++ client/coral-embed-stream/src/reducers/index.js
@@ -1,5 +1,4 @@
 import embed from './embed';
-import comment from './comment';
 import config from './config';
 import stream from './stream';
 
@@ -7,5 +6,4 @@ export default {
   embed,
   config,
   stream,
-  comment
 };
