+++ plugin-api/alpha/client/actions/index.js
@@ -1,2 +1,2 @@
 export {addTag, removeTag} from 'coral-plugin-commentbox/actions';
-export {addCommentClassName, removeCommentClassName} from 'coral-embed-stream/src/actions/comment';
+export {addCommentClassName, removeCommentClassName} from 'coral-embed-stream/src/actions/stream';
