+++ plugins/coral-plugin-offtopic/client/containers/OffTopicCheckbox.js
@@ -1,9 +1,12 @@
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 import {addTag, removeTag} from 'plugin-api/alpha/client/actions';
+import {commentBoxTagsSelector} from 'plugin-api/alpha/client/selectors';
 import OffTopicCheckbox from '../components/OffTopicCheckbox';
 
-const mapStateToProps = ({commentBox}) => ({commentBox});
+const mapStateToProps = (state) => ({
+  tags: commentBoxTagsSelector(state)
+});
 
 const mapDispatchToProps = (dispatch) =>
   bindActionCreators({addTag, removeTag}, dispatch);
